--- frontend_app/pages/course/components/LocationCard.vue
@@ -1,97 +1,0 @@
-<template>
-    <div style="margin: 20px 40px;">
-        <div v-for="location in locations" :key="location.value" class="location-item">
-            {{ location.value }}: {{ location.label }}
-            <el-icon :size="14" @click.stop="$emit('delete-location', location.value)" class="location-icon">
-                <Close />
-            </el-icon>
-        </div>
-
-        <div class="form-container">
-            <h1 style="margin-bottom: 10px;">Create new Location</h1>
-            <el-form ref="locationForm" :model="locationForm" label-width="80px">
-                <el-form-item label="Name">
-                    <el-input v-model="locationForm.name"></el-input>
-                </el-form-item>
-                <el-form-item>
-                    <el-button type="primary" @click="submitLocation">Submit Location</el-button>
-                    <p style="font-size: small; color: darkgrey;">( Your current location will be used )</p>
-                </el-form-item>
-            </el-form>
-        </div>
-
-    </div>
-</template>
-  
-<script>
-export default {
-    emits: ['create-location', 'delete-location'],
-    props: {
-        locations: Array
-    },
-    data() {
-        return {
-            locationForm: {
-                name: '',
-                latitude: '',
-                longitude: ''
-            }
-        }
-    },
-    watch: {
-
-    },
-    methods: {
-        getLocation() {
-            // Check if Geolocation is supported
-            if ("geolocation" in navigator) {
-                navigator.geolocation.getCurrentPosition((position) => {
-                    const { latitude, longitude } = position.coords
-                    const locationData = {
-                        name: this.locationForm.name,
-                        latitude: latitude,
-                        longitude: longitude
-                    };
-                    this.$emit('create-location', locationData)
-                    this.resetForm()
-                }, (error) => {
-                    // Handle location 
-                    console.error('Error getting location', error);
-                });
-            } else {
-                // Geolocation is not supported by this browser
-                console.error('Geolocation is not supported by this browser.');
-            }
-        },
-        submitLocation() {
-            if (this.locationForm.name.trim() === '') {
-                alert('Please enter a location name.');
-                return;
-            }
-            this.getLocation(); // Get location coords then create new location
-        },
-        resetForm() {
-            this.locationForm = { name: '', latitude: '', longitude: '' };
-        }
-    }
-}
-</script>
-<style scoped>
-.form-container {
-    width: 400px;
-    margin: 30px 20px;
-}
-
-.location-item {
-    display: flex;
-    align-items: center;
-    /* Ensures vertical alignment in case icon and text have different heights */
-}
-
-.location-icon {
-    cursor: pointer;
-    /* Changes the cursor to a pointer when hovering over the icon */
-    margin-left: 5px;
-    /* Optional: Adds a small space between the text and the icon */
-}
-</style>
