--- frontend_app/pages/course/index.vue
@@ -1,16 +1,0 @@
-<template>
-    <router-view></router-view> 
-</template>
-
-<script>
-  export default {
-    name: 'Course',
-    data() {
-        return {
-        }
-    },
-    created() {
- }
-  }
-
-</script>
