+++ frontend_app/router/index.js
@@ -1,59 +1,17 @@
-import { createRouter, createWebHashHistory } from 'vue-router';
-import LoginPage from '../pages/Login.vue';
-import ManageAccount from '../pages/ManageAccount.vue';
-import Course from '../pages/course/index.vue';
-import AllCourses from '../pages/course/AllCourse.vue'
-import SingleCourse from '../pages/course/SingleCourse.vue'
-import ManageCourse from '../pages/ManageCourse.vue';
-import PageNotFound from '../pages/404.vue'
-import AttendanceTrack from '../pages/course/AttendanceTrack.vue'
+import { createRouter, createMemoryHistory, createWebHistory } from 'vue-router';
+import Dashboard from '../pages/Dashboard.vue'
+
 
 const routes = [
   {
-    path: '/login',
-    name: 'Login',
-    component: LoginPage
+    path: '/',
+    name: 'Home',
+    component: Dashboard,
   },
-  {
-    path: '/course',
-    name: 'Course',
-    component: Course,
-    children: [
-      {
-        path: '',
-        name: 'Courses',
-        component: AllCourses,
-      },
-      {
-        path: ':id',
-        name: 'SingleCourse',
-        component: SingleCourse
-      },
-      {
-        path: ':id/attendance',
-        name: 'AttendanceTrack',
-        component: AttendanceTrack
-      },
-    ]
-  },
-  {
-    path: '/manage-account',
-    name: 'ManageAccount',
-    component: ManageAccount
-  },
-  {
-    path: '/manage-course',
-    name: 'ManageCourse',
-    component: ManageCourse
-  },
-  {
-    path: "/:notFound",
-    component: PageNotFound,
-  }
-]
+];
 
 const router = createRouter({
-  history: createWebHashHistory(),
+  history: createWebHistory(),
   routes
 })
